centos/*/*-manage.py: use setuid instead of seteuid
Firefox (run through selenium) was switching back to root user with
seteuid() -- breaking tests.

There seems to be no need for the manage script to regain root
privileges, so I use setuid() call instead of seteuid().

Other problem with this user changing business is that some python
modules such as ipython use the HOME env variable which is often
/root. So I set the HOME environment variable to apache's home
directory. However this still isn't good because apache can't
necessarily write to its own home directory.

It could be nice to create a registry user with the RPM and have
mod_wsgi run with this user. Then the manage script could refuse to
work unless running as the registry user.